backend/requirements.txt
Flask
web3
python-dotenv
Flask-Cors
